Add checkpoint/recovery mechanism 1. make a checkpoint before create subrequest, including pool,pool->large(malloc),pool cleanup handler,    http cleanup handler, chain. 2. recovery to last checkpoint after finish the subrequest(before create the next subrequest)
We do this to release files and free memory brought in by subrequest so that we can keep a
small counts of opened fd ,and small memory footprint.
This is important for subrange a very large file (>1G) to reduce resouce occupation by subrequest,because
the resouce will never be used later.